Decompose codegen.py into codegen/ mixin package
Split the 2,140-line codegen.py monolith into 11 focused modules
following the mixin pattern established by checker/ and wasm/:

  codegen/api.py          - Public API, compile(), execute()
  codegen/core.py         - Composed CodeGenerator class
  codegen/modules.py      - Cross-module call detection (Pass 0)
  codegen/registration.py - Function/ADT registration (Pass 1)
  codegen/monomorphize.py - Generic instantiation (Pass 1.5)
  codegen/functions.py    - Function body compilation (Pass 2)
  codegen/closures.py     - Closure lifting
  codegen/contracts.py    - Runtime contract insertion
  codegen/assembly.py     - WAT module assembly
  codegen/compilability.py - Compilability checks

Also adds 5 coverage gap tests (E600/E601/E605/E606, unknown module
calls) and splits the 4,834-line test_codegen.py into 6 focused files:
test_codegen_contracts.py, test_codegen_monomorphize.py,
test_codegen_closures.py, test_codegen_modules.py, and
test_codegen_coverage.py.

All 1,076 tests pass, mypy clean, all 14 examples validate.

Closes #155
