+++ tests/test_codegen.py
@@ -381,45 +381,6 @@ class TestCompileResult:
         assert result.ok is True
 
 
-class TestCrossModuleGuardRail:
-    """Calls to undefined functions produce a proper diagnostic."""
-
-    def test_undefined_fn_call_diagnostic(self) -> None:
-        """Calling a function not defined in this module emits a diagnostic."""
-        source = """\
-public fn f(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ unknown_fn(42) }
-"""
-        result = _compile(source)
-        errors = [d for d in result.diagnostics if d.severity == "error"]
-        assert len(errors) == 1
-        assert "unknown_fn" in errors[0].description
-        assert "not defined in this module" in errors[0].description
-        assert "not found in any imported module" in errors[0].description
-        assert result.ok is False
-
-    def test_undefined_fn_no_raw_wasmtime_error(self) -> None:
-        """No raw WAT compilation error — guard rail catches it first."""
-        source = """\
-public fn f(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ missing(1) }
-"""
-        result = _compile(source)
-        errors = [d for d in result.diagnostics if d.severity == "error"]
-        assert all("WAT compilation failed" not in e.description for e in errors)
-
-    def test_locally_defined_fn_compiles(self) -> None:
-        """Calls to locally defined functions still work."""
-        source = """\
-public fn helper(-> @Int) requires(true) ensures(true) effects(pure) { 1 }
-public fn f(-> @Int) requires(true) ensures(true) effects(pure) { helper() }
-"""
-        result = _compile_ok(source)
-        assert result.ok is True
-
-
 # =====================================================================
 # 5b: Slot references + arithmetic
 # =====================================================================
@@ -960,206 +921,6 @@ public fn main(@Unit -> @Unit)
         assert result.ok
         exec_result = execute(result, fn_name="main")
         assert exec_result.stdout == "Hello, World!"
-
-
-# =====================================================================
-# 5f: Runtime contract insertion
-# =====================================================================
-
-
-class TestPreconditions:
-    def test_requires_holds(self) -> None:
-        """Non-trivial precondition that holds — no trap."""
-        source = """\
-public fn positive(@Int -> @Int)
-  requires(@Int.0 > 0)
-  ensures(true)
-  effects(pure)
-{ @Int.0 }
-"""
-        assert _run(source, fn="positive", args=[5]) == 5
-
-    def test_requires_traps(self) -> None:
-        """Non-trivial precondition violated — WASM trap."""
-        source = """\
-public fn positive(@Int -> @Int)
-  requires(@Int.0 > 0)
-  ensures(true)
-  effects(pure)
-{ @Int.0 }
-"""
-        _run_trap(source, fn="positive", args=[0])
-
-    def test_requires_boundary(self) -> None:
-        """Precondition with exact boundary value."""
-        source = """\
-public fn nonneg(@Int -> @Int)
-  requires(@Int.0 >= 0)
-  ensures(true)
-  effects(pure)
-{ @Int.0 }
-"""
-        assert _run(source, fn="nonneg", args=[0]) == 0
-        _run_trap(source, fn="nonneg", args=[-1])
-
-    def test_requires_neq_zero(self) -> None:
-        """Precondition: denominator != 0."""
-        source = """\
-public fn safe_div(@Int, @Int -> @Int)
-  requires(@Int.0 != 0)
-  ensures(true)
-  effects(pure)
-{ @Int.1 / @Int.0 }
-"""
-        assert _run(source, fn="safe_div", args=[10, 2]) == 5
-        _run_trap(source, fn="safe_div", args=[10, 0])
-
-    def test_trivial_requires_no_overhead(self) -> None:
-        """requires(true) should not produce any trap instructions."""
-        source = """\
-public fn f(@Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{ @Int.0 }
-"""
-        result = _compile_ok(source)
-        # No unreachable in WAT (no contract checks needed)
-        assert "unreachable" not in result.wat
-
-    def test_multiple_requires(self) -> None:
-        """Multiple preconditions — all must hold."""
-        source = """\
-public fn bounded(@Int -> @Int)
-  requires(@Int.0 >= 0)
-  requires(@Int.0 <= 100)
-  ensures(true)
-  effects(pure)
-{ @Int.0 }
-"""
-        assert _run(source, fn="bounded", args=[50]) == 50
-        _run_trap(source, fn="bounded", args=[-1])
-        _run_trap(source, fn="bounded", args=[101])
-
-
-class TestPostconditions:
-    def test_ensures_holds(self) -> None:
-        """Postcondition that holds — no trap."""
-        source = """\
-public fn double(@Int -> @Int)
-  requires(true)
-  ensures(@Int.result >= 0)
-  effects(pure)
-{ @Int.0 * 2 }
-"""
-        assert _run(source, fn="double", args=[5]) == 10
-
-    def test_ensures_traps(self) -> None:
-        """Postcondition violated — WASM trap."""
-        source = """\
-public fn negate(@Int -> @Int)
-  requires(true)
-  ensures(@Int.result > 0)
-  effects(pure)
-{ -@Int.0 }
-"""
-        # negate(5) returns -5, which violates ensures(result > 0)
-        _run_trap(source, fn="negate", args=[5])
-
-    def test_ensures_with_params(self) -> None:
-        """Postcondition referencing both result and parameters."""
-        source = """\
-public fn inc(@Int -> @Int)
-  requires(true)
-  ensures(@Int.result > @Int.0)
-  effects(pure)
-{ @Int.0 + 1 }
-"""
-        assert _run(source, fn="inc", args=[5]) == 6
-
-    def test_ensures_result_eq(self) -> None:
-        """Postcondition checking exact result value."""
-        source = """\
-public fn always_zero(-> @Int)
-  requires(true)
-  ensures(@Int.result == 0)
-  effects(pure)
-{ 0 }
-"""
-        assert _run(source, fn="always_zero") == 0
-
-    def test_ensures_result_traps(self) -> None:
-        """Postcondition checking exact value — wrong result traps."""
-        source = """\
-public fn buggy(-> @Int)
-  requires(true)
-  ensures(@Int.result == 0)
-  effects(pure)
-{ 42 }
-"""
-        _run_trap(source, fn="buggy")
-
-    def test_trivial_ensures_no_overhead(self) -> None:
-        """ensures(true) should not produce any trap instructions."""
-        source = """\
-public fn f(@Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{ @Int.0 }
-"""
-        result = _compile_ok(source)
-        assert "unreachable" not in result.wat
-
-    def test_ensures_bool_result(self) -> None:
-        """Postcondition on a Bool-returning function."""
-        source = """\
-public fn is_pos(@Int -> @Bool)
-  requires(true)
-  ensures(@Bool.result == true)
-  effects(pure)
-{ @Int.0 > 0 }
-"""
-        assert _run(source, fn="is_pos", args=[5]) == 1
-        # is_pos(-1) returns false, violating ensures(result == true)
-        _run_trap(source, fn="is_pos", args=[-1])
-
-
-class TestCombinedContracts:
-    def test_both_hold(self) -> None:
-        """Both requires and ensures hold — normal execution."""
-        source = """\
-public fn safe_inc(@Int -> @Int)
-  requires(@Int.0 >= 0)
-  ensures(@Int.result > @Int.0)
-  effects(pure)
-{ @Int.0 + 1 }
-"""
-        assert _run(source, fn="safe_inc", args=[0]) == 1
-        assert _run(source, fn="safe_inc", args=[10]) == 11
-
-    def test_requires_fails_first(self) -> None:
-        """Precondition fails before postcondition is checked."""
-        source = """\
-public fn safe_inc(@Int -> @Int)
-  requires(@Int.0 >= 0)
-  ensures(@Int.result > @Int.0)
-  effects(pure)
-{ @Int.0 + 1 }
-"""
-        _run_trap(source, fn="safe_inc", args=[-1])
-
-    def test_contracts_with_recursion(self) -> None:
-        """Runtime contracts on a recursive function."""
-        source = """\
-public fn factorial(@Nat -> @Nat)
-  requires(@Nat.0 >= 0)
-  ensures(@Nat.result >= 1)
-  decreases(@Nat.0)
-  effects(pure)
-{
-  if @Nat.0 <= 1 then { 1 }
-  else { @Nat.0 * factorial(@Nat.0 - 1) }
-}
-"""
-        assert _run(source, fn="factorial", args=[5]) == 120
-        assert _run(source, fn="factorial", args=[0]) == 1
 
 
 # =====================================================================
@@ -2539,600 +2300,6 @@ public fn unwrap_or(@Option<Int> -> @Int)
 
 
 # =====================================================================
-# C6i: Monomorphization of generic (forall<T>) functions
-# =====================================================================
-
-
-class TestMonomorphization:
-    """Tests for monomorphization of forall<T> functions."""
-
-    def test_identity_int(self) -> None:
-        """forall<T> fn identity instantiated with Int."""
-        source = """\
-private forall<T> fn identity(@T -> @T)
-  requires(true) ensures(true) effects(pure)
-{ @T.0 }
-
-public fn main(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ identity(42) }
-"""
-        assert _run(source, fn="main") == 42
-
-    def test_identity_bool(self) -> None:
-        """forall<T> fn identity instantiated with Bool."""
-        source = """\
-private forall<T> fn identity(@T -> @T)
-  requires(true) ensures(true) effects(pure)
-{ @T.0 }
-
-public fn main(-> @Bool)
-  requires(true) ensures(true) effects(pure)
-{ identity(true) }
-"""
-        assert _run(source, fn="main") == 1
-
-    def test_identity_two_instantiations(self) -> None:
-        """Same generic function instantiated with both Int and Bool."""
-        source = """\
-private forall<T> fn identity(@T -> @T)
-  requires(true) ensures(true) effects(pure)
-{ @T.0 }
-
-public fn test_int(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ identity(42) }
-
-public fn test_bool(-> @Bool)
-  requires(true) ensures(true) effects(pure)
-{ identity(false) }
-"""
-        result = _compile_ok(source)
-        # Private generic → monomorphized variants not exported
-        assert "identity$Int" not in result.exports
-        assert "identity$Bool" not in result.exports
-        # Public callers are exported
-        assert "test_int" in result.exports
-        assert "test_bool" in result.exports
-        # Run both
-        exec_int = execute(result, fn_name="test_int")
-        assert exec_int.value == 42
-        exec_bool = execute(result, fn_name="test_bool")
-        assert exec_bool.value == 0
-
-    def test_identity_slot_ref_arg(self) -> None:
-        """Generic function called with a slot reference argument."""
-        source = """\
-private forall<T> fn identity(@T -> @T)
-  requires(true) ensures(true) effects(pure)
-{ @T.0 }
-
-public fn main(@Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{ identity(@Int.0) }
-"""
-        assert _run(source, fn="main", args=[99]) == 99
-
-    def test_const_function(self) -> None:
-        """forall<A, B> fn const with two type parameters."""
-        source = """\
-private forall<A, B> fn const(@A, @B -> @A)
-  requires(true) ensures(true) effects(pure)
-{ @A.0 }
-
-public fn main(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ const(42, true) }
-"""
-        assert _run(source, fn="main") == 42
-
-    def test_generic_with_adt_match(self) -> None:
-        """forall<T> fn is_some with ADT match (Some case)."""
-        source = """\
-private data Option<T> { None, Some(T) }
-
-private forall<T> fn is_some(@Option<T> -> @Bool)
-  requires(true) ensures(true) effects(pure)
-{
-  match @Option<T>.0 {
-    None -> false,
-    Some(@T) -> true
-  }
-}
-
-public fn main(-> @Bool)
-  requires(true) ensures(true) effects(pure)
-{ is_some(Some(1)) }
-"""
-        assert _run(source, fn="main") == 1
-
-    def test_generic_with_adt_match_none(self) -> None:
-        """forall<T> fn is_some with ADT match (None case)."""
-        source = """\
-private data Option<T> { None, Some(T) }
-
-private forall<T> fn is_some(@Option<T> -> @Bool)
-  requires(true) ensures(true) effects(pure)
-{
-  match @Option<T>.0 {
-    None -> false,
-    Some(@T) -> true
-  }
-}
-
-public fn main(-> @Bool)
-  requires(true) ensures(true) effects(pure)
-{
-  let @Option<Int> = None;
-  is_some(@Option<Int>.0)
-}
-"""
-        assert _run(source, fn="main") == 0
-
-    def test_generic_fn_wat_has_mangled_name(self) -> None:
-        """WAT output contains mangled function name."""
-        source = """\
-private forall<T> fn identity(@T -> @T)
-  requires(true) ensures(true) effects(pure)
-{ @T.0 }
-
-public fn main(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ identity(42) }
-"""
-        result = _compile_ok(source)
-        assert "$identity$Int" in result.wat
-
-    def test_generic_fn_mangled_in_exports(self) -> None:
-        """Private generic's mangled names not exported; public caller is."""
-        source = """\
-private forall<T> fn identity(@T -> @T)
-  requires(true) ensures(true) effects(pure)
-{ @T.0 }
-
-public fn main(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ identity(42) }
-"""
-        result = _compile_ok(source)
-        # Private generic → monomorphized variants not exported
-        assert "identity$Int" not in result.exports
-        assert "identity" not in result.exports
-        # Public caller is exported
-        assert "main" in result.exports
-
-    def test_non_generic_fn_unaffected(self) -> None:
-        """Non-generic functions compile normally alongside generic ones."""
-        source = """\
-private forall<T> fn identity(@T -> @T)
-  requires(true) ensures(true) effects(pure)
-{ @T.0 }
-
-public fn double(@Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{ @Int.0 + @Int.0 }
-
-public fn main(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ double(identity(21)) }
-"""
-        assert _run(source, fn="main") == 42
-
-    def test_generic_identity_in_let_binding(self) -> None:
-        """Generic call result used in a let binding."""
-        source = """\
-private forall<T> fn identity(@T -> @T)
-  requires(true) ensures(true) effects(pure)
-{ @T.0 }
-
-public fn main(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  let @Int = identity(10);
-  @Int.0 + 5
-}
-"""
-        assert _run(source, fn="main") == 15
-
-    def test_generic_chained_calls(self) -> None:
-        """Generic function called with result of another generic call."""
-        source = """\
-private forall<T> fn identity(@T -> @T)
-  requires(true) ensures(true) effects(pure)
-{ @T.0 }
-
-public fn main(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ identity(identity(99)) }
-"""
-        assert _run(source, fn="main") == 99
-
-    def test_generic_in_if_branch(self) -> None:
-        """Generic call inside an if-then-else branch."""
-        source = """\
-private forall<T> fn identity(@T -> @T)
-  requires(true) ensures(true) effects(pure)
-{ @T.0 }
-
-public fn main(@Bool -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  if @Bool.0 then { identity(1) } else { identity(2) }
-}
-"""
-        assert _run(source, fn="main", args=[1]) == 1
-        assert _run(source, fn="main", args=[0]) == 2
-
-    def test_generic_with_arithmetic_arg(self) -> None:
-        """Generic function called with arithmetic expression as argument."""
-        source = """\
-private forall<T> fn identity(@T -> @T)
-  requires(true) ensures(true) effects(pure)
-{ @T.0 }
-
-public fn main(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ identity(3 + 4) }
-"""
-        assert _run(source, fn="main") == 7
-
-    def test_generic_no_callers_skipped(self) -> None:
-        """Generic function with no callers is gracefully skipped."""
-        source = """\
-private forall<T> fn identity(@T -> @T)
-  requires(true) ensures(true) effects(pure)
-{ @T.0 }
-
-public fn main(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ 42 }
-"""
-        result = _compile_ok(source)
-        assert "main" in result.exports
-        # identity has no callers → no monomorphized version → not in exports
-        assert "identity" not in result.exports
-
-    def test_generics_example_file(self) -> None:
-        """examples/generics.vera compiles without errors."""
-        from pathlib import Path
-        path = Path(__file__).parent.parent / "examples" / "generics.vera"
-        source = path.read_text()
-        result = _compile(source)
-        assert result.ok
-
-    def test_list_ops_example_file(self) -> None:
-        """examples/list_ops.vera still compiles (concrete, not generic)."""
-        from pathlib import Path
-        path = Path(__file__).parent.parent / "examples" / "list_ops.vera"
-        source = path.read_text()
-        result = _compile(source)
-        assert result.ok
-
-
-# =====================================================================
-# C6h: Closures
-# =====================================================================
-
-
-class TestClosures:
-    """Tests for closure compilation — anonymous functions, captures,
-    apply_fn, function tables, and call_indirect."""
-
-    def test_closure_no_capture(self) -> None:
-        """An anonymous function with no free variables compiles and runs."""
-        src = """\
-type IntToInt = fn(Int -> Int) effects(pure);
-public fn make_fn(@Unit -> @IntToInt)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 * 2 }
-}
-public fn test(@Unit -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  let @IntToInt = make_fn(());
-  apply_fn(@IntToInt.0, 7)
-}
-"""
-        assert _run(src, "test") == 14
-
-    def test_closure_with_capture(self) -> None:
-        """An anonymous function that captures an outer binding."""
-        src = """\
-type IntToInt = fn(Int -> Int) effects(pure);
-public fn make_adder(@Int -> @IntToInt)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 + @Int.1 }
-}
-public fn test(@Unit -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  let @IntToInt = make_adder(10);
-  apply_fn(@IntToInt.0, 5)
-}
-"""
-        assert _run(src, "test") == 15
-
-    def test_apply_fn_basic(self) -> None:
-        """apply_fn invokes a closure with the correct argument."""
-        src = """\
-type IntToInt = fn(Int -> Int) effects(pure);
-public fn make_doubler(@Unit -> @IntToInt)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 * 2 }
-}
-public fn test(@Unit -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  let @IntToInt = make_doubler(());
-  apply_fn(@IntToInt.0, 21)
-}
-"""
-        assert _run(src, "test") == 42
-
-    def test_apply_fn_with_capture(self) -> None:
-        """apply_fn on a capturing closure produces the correct result."""
-        src = """\
-type IntToInt = fn(Int -> Int) effects(pure);
-public fn make_multiplier(@Int -> @IntToInt)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 * @Int.1 }
-}
-public fn test(@Unit -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  let @IntToInt = make_multiplier(3);
-  apply_fn(@IntToInt.0, 7)
-}
-"""
-        assert _run(src, "test") == 21
-
-    def test_closure_in_let(self) -> None:
-        """Store a closure in a let binding, then use it."""
-        src = """\
-type IntToInt = fn(Int -> Int) effects(pure);
-public fn make_fn(@Int -> @IntToInt)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 + @Int.1 }
-}
-public fn test(@Unit -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  let @IntToInt = make_fn(100);
-  let @Int = apply_fn(@IntToInt.0, 23);
-  @Int.0
-}
-"""
-        assert _run(src, "test") == 123
-
-    def test_closure_as_param(self) -> None:
-        """Pass a closure as a function parameter."""
-        src = """\
-type IntToInt = fn(Int -> Int) effects(pure);
-public fn apply(@IntToInt, @Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  apply_fn(@IntToInt.0, @Int.0)
-}
-public fn make_fn(@Int -> @IntToInt)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 + @Int.1 }
-}
-public fn test(@Unit -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  let @IntToInt = make_fn(50);
-  apply(@IntToInt.0, 50)
-}
-"""
-        assert _run(src, "test") == 100
-
-    def test_closure_in_match(self) -> None:
-        """Use a closure inside a match arm with an ADT constructor."""
-        src = """\
-private data Option<T> { None, Some(T) }
-type IntMapper = fn(Int -> Int) effects(pure);
-public fn map_option(@Option<Int>, @IntMapper -> @Option<Int>)
-  requires(true) ensures(true) effects(pure)
-{
-  match @Option<Int>.0 {
-    None -> None,
-    Some(@Int) -> Some(apply_fn(@IntMapper.0, @Int.0))
-  }
-}
-public fn make_adder(@Int -> @IntMapper)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 + @Int.1 }
-}
-public fn test(@Unit -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  let @IntMapper = make_adder(100);
-  let @Option<Int> = map_option(Some(5), @IntMapper.0);
-  match @Option<Int>.0 {
-    None -> 0,
-    Some(@Int) -> @Int.0
-  }
-}
-"""
-        assert _run(src, "test") == 105
-
-    def test_closure_in_match_none(self) -> None:
-        """map_option on None returns None."""
-        src = """\
-private data Option<T> { None, Some(T) }
-type IntMapper = fn(Int -> Int) effects(pure);
-public fn map_option(@Option<Int>, @IntMapper -> @Option<Int>)
-  requires(true) ensures(true) effects(pure)
-{
-  match @Option<Int>.0 {
-    None -> None,
-    Some(@Int) -> Some(apply_fn(@IntMapper.0, @Int.0))
-  }
-}
-public fn make_adder(@Int -> @IntMapper)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 + @Int.1 }
-}
-public fn test(@Unit -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  let @IntMapper = make_adder(100);
-  let @Option<Int> = map_option(None, @IntMapper.0);
-  match @Option<Int>.0 {
-    None -> -1,
-    Some(@Int) -> @Int.0
-  }
-}
-"""
-        assert _run(src, "test") == -1
-
-    def test_fn_type_param_compiles(self) -> None:
-        """A function with a function-type parameter is not skipped."""
-        src = """\
-type IntToInt = fn(Int -> Int) effects(pure);
-public fn apply(@IntToInt, @Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  apply_fn(@IntToInt.0, @Int.0)
-}
-"""
-        result = _compile_ok(src)
-        assert "apply" in result.exports
-
-    def test_table_in_wat(self) -> None:
-        """WAT output includes a funcref table when closures are used."""
-        src = """\
-type IntToInt = fn(Int -> Int) effects(pure);
-public fn make_fn(@Unit -> @IntToInt)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 }
-}
-"""
-        result = _compile_ok(src)
-        assert result.wat is not None
-        assert "funcref" in result.wat
-        assert "(table" in result.wat
-        assert "(elem" in result.wat
-
-    def test_call_indirect_in_wat(self) -> None:
-        """WAT output contains call_indirect for apply_fn."""
-        src = """\
-type IntToInt = fn(Int -> Int) effects(pure);
-public fn apply(@IntToInt, @Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  apply_fn(@IntToInt.0, @Int.0)
-}
-public fn make_fn(@Unit -> @IntToInt)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 }
-}
-public fn test(@Unit -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  let @IntToInt = make_fn(());
-  apply(@IntToInt.0, 99)
-}
-"""
-        result = _compile_ok(src)
-        assert result.wat is not None
-        assert "call_indirect" in result.wat
-
-    def test_type_sig_in_wat(self) -> None:
-        """WAT output contains a closure type signature declaration."""
-        src = """\
-type IntToInt = fn(Int -> Int) effects(pure);
-public fn make_fn(@Unit -> @IntToInt)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 }
-}
-"""
-        result = _compile_ok(src)
-        assert result.wat is not None
-        assert "$closure_sig_" in result.wat
-        assert "(type" in result.wat
-
-    def test_closures_example_compiles(self) -> None:
-        """examples/closures.vera compiles without errors."""
-        from pathlib import Path
-        path = Path(__file__).parent.parent / "examples" / "closures.vera"
-        source = path.read_text()
-        result = _compile(source)
-        assert result.ok
-
-    def test_closures_example_test_closure(self) -> None:
-        """examples/closures.vera test_closure returns 15."""
-        from pathlib import Path
-        path = Path(__file__).parent.parent / "examples" / "closures.vera"
-        source = path.read_text()
-        result = _compile_ok(source)
-        exec_result = execute(result, fn_name="test_closure")
-        assert exec_result.value == 15
-
-    def test_closures_example_test_map_option(self) -> None:
-        """examples/closures.vera test_map_option returns 105."""
-        from pathlib import Path
-        path = Path(__file__).parent.parent / "examples" / "closures.vera"
-        source = path.read_text()
-        result = _compile_ok(source)
-        exec_result = execute(result, fn_name="test_map_option")
-        assert exec_result.value == 105
-
-    def test_multiple_closures(self) -> None:
-        """Multiple closures get distinct table entries."""
-        src = """\
-type IntToInt = fn(Int -> Int) effects(pure);
-public fn make_adder(@Int -> @IntToInt)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 + @Int.1 }
-}
-public fn test(@Unit -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  let @IntToInt = make_adder(10);
-  let @Int = apply_fn(@IntToInt.0, 5);
-  let @IntToInt = make_adder(20);
-  let @Int = apply_fn(@IntToInt.0, 3);
-  @Int.0 + @Int.1
-}
-"""
-        assert _run(src, "test") == 38  # 15 + 23
-
-    def test_closure_captures_correct_value(self) -> None:
-        """Each closure captures the value at its creation point."""
-        src = """\
-type IntToInt = fn(Int -> Int) effects(pure);
-public fn make_adder(@Int -> @IntToInt)
-  requires(true) ensures(true) effects(pure)
-{
-  fn(@Int -> @Int) effects(pure) { @Int.0 + @Int.1 }
-}
-public fn test(@Unit -> @Int)
-  requires(true) ensures(true) effects(pure)
-{
-  let @IntToInt = make_adder(1);
-  let @Int = apply_fn(@IntToInt.0, 0);
-  let @IntToInt = make_adder(100);
-  let @Int = apply_fn(@IntToInt.0, 0);
-  @Int.0 + @Int.1
-}
-"""
-        assert _run(src, "test") == 101  # 1 + 100
-
-
-# =====================================================================
 # C6j: Effect Handlers
 # =====================================================================
 
@@ -4107,501 +3274,6 @@ public fn hello(-> @String)
         exec_result = execute(result, fn_name="hello")
         # Returns the data pointer (an integer)
         assert isinstance(exec_result.value, int)
-
-
-# =====================================================================
-# 6.5f: old()/new() state expressions in postconditions
-# =====================================================================
-
-
-class TestOldNewContracts:
-    """Tests for old()/new() state expression compilation in postconditions."""
-
-    def test_old_new_postcondition_compiles(self) -> None:
-        """Function with old()/new() in ensures clause compiles to WASM."""
-        src = """
-public fn increment(@Unit -> @Unit)
-  requires(true)
-  ensures(new(State<Int>) == old(State<Int>) + 1)
-  effects(<State<Int>>)
-{
-  let @Int = get(());
-  put(@Int.0 + 1);
-  ()
-}
-"""
-        result = _compile_ok(src)
-        assert "increment" in result.exports
-
-    def test_old_new_postcondition_passes(self) -> None:
-        """Postcondition holds — no trap when new == old + 1."""
-        src = """
-public fn increment(@Unit -> @Unit)
-  requires(true)
-  ensures(new(State<Int>) == old(State<Int>) + 1)
-  effects(<State<Int>>)
-{
-  let @Int = get(());
-  put(@Int.0 + 1);
-  ()
-}
-"""
-        result = _compile_ok(src)
-        exec_result = execute(
-            result, fn_name="increment",
-            initial_state={"State_Int": 10},
-        )
-        # Should complete without trap
-        assert exec_result.value is None  # Unit return
-
-    def test_old_new_postcondition_traps(self) -> None:
-        """Postcondition violated — traps when increment is wrong."""
-        src = """
-public fn bad_increment(@Unit -> @Unit)
-  requires(true)
-  ensures(new(State<Int>) == old(State<Int>) + 1)
-  effects(<State<Int>>)
-{
-  let @Int = get(());
-  put(@Int.0 + 2);
-  ()
-}
-"""
-        result = _compile_ok(src)
-        with pytest.raises((wasmtime.WasmtimeError, wasmtime.Trap, RuntimeError)):
-            execute(
-                result, fn_name="bad_increment",
-                initial_state={"State_Int": 5},
-            )
-
-    def test_trivial_ensures_no_snapshot(self) -> None:
-        """ensures(true) with State effect does NOT emit a snapshot."""
-        src = """
-public fn inc(@Unit -> @Unit)
-  requires(true) ensures(true)
-  effects(<State<Int>>)
-{
-  let @Int = get(());
-  put(@Int.0 + 1);
-  ()
-}
-"""
-        result = _compile_ok(src)
-        wat = result.wat
-        assert "inc" in result.exports
-        # Body should call state_get for the let binding,
-        # but no snapshot local.set before the body
-        lines = wat.split("\n")
-        # Find the function body — there should be exactly one
-        # state_get call (the let binding), not two (snapshot + let)
-        state_get_count = sum(
-            1 for l in lines if "call $vera.state_get_Int" in l
-        )
-        # Only the body's get() call — no snapshot
-        assert state_get_count == 1
-
-    def test_old_new_wat_structure(self) -> None:
-        """WAT contains state_get snapshot before body and new() in postcondition."""
-        src = """
-public fn increment(@Unit -> @Unit)
-  requires(true)
-  ensures(new(State<Int>) == old(State<Int>) + 1)
-  effects(<State<Int>>)
-{
-  let @Int = get(());
-  put(@Int.0 + 1);
-  ()
-}
-"""
-        result = _compile_ok(src)
-        wat = result.wat
-        lines = wat.split("\n")
-        state_get_count = sum(
-            1 for l in lines if "call $vera.state_get_Int" in l
-        )
-        # 3 calls: snapshot (old), body get(), postcondition new()
-        assert state_get_count == 3
-
-    def test_new_reads_current_state(self) -> None:
-        """new(State<T>) reads the current value, not the snapshot."""
-        # Increment by 5 but claim increment by 5 in postcondition
-        src = """
-public fn add_five(@Unit -> @Unit)
-  requires(true)
-  ensures(new(State<Int>) == old(State<Int>) + 5)
-  effects(<State<Int>>)
-{
-  let @Int = get(());
-  put(@Int.0 + 5);
-  ()
-}
-"""
-        result = _compile_ok(src)
-        exec_result = execute(
-            result, fn_name="add_five",
-            initial_state={"State_Int": 100},
-        )
-        assert exec_result.value is None  # Unit, no trap
-
-
-# =====================================================================
-# Cross-module codegen (C7e)
-# =====================================================================
-
-
-class TestCrossModuleCodegen:
-    """Imported functions are compiled into the WASM module via flattening."""
-
-    # Reusable module sources
-    MATH_MODULE = """\
-public fn abs(@Int -> @Int)
-  requires(true)
-  ensures(@Int.result >= 0)
-  effects(pure)
-{ if @Int.0 < 0 then { 0 - @Int.0 } else { @Int.0 } }
-
-public fn max(@Int, @Int -> @Int)
-  requires(true)
-  ensures(@Int.result >= @Int.0)
-  ensures(@Int.result >= @Int.1)
-  effects(pure)
-{ if @Int.0 >= @Int.1 then { @Int.0 } else { @Int.1 } }
-"""
-
-    HELPER_MODULE = """\
-public fn double(@Int -> @Int)
-  requires(true)
-  ensures(true)
-  effects(pure)
-{ internal(@Int.0) + internal(@Int.0) }
-
-private fn internal(@Int -> @Int)
-  requires(true)
-  ensures(true)
-  effects(pure)
-{ @Int.0 }
-"""
-
-    @staticmethod
-    def _resolved(
-        path: tuple[str, ...], source: str,
-    ) -> "ResolvedModule":
-        """Build a ResolvedModule from source text."""
-        import tempfile
-        from pathlib import Path
-
-        from vera.resolver import ResolvedModule as RM
-
-        with tempfile.NamedTemporaryFile(
-            mode="w", suffix=".vera", delete=False
-        ) as f:
-            f.write(source)
-            f.flush()
-            fpath = f.name
-
-        tree = parse_file(fpath)
-        prog = transform(tree)
-        return RM(
-            path=path,
-            file_path=Path(fpath),
-            program=prog,
-            source=source,
-        )
-
-    @classmethod
-    def _compile_mod(
-        cls, source: str, modules: list,
-    ) -> CompileResult:
-        """Compile with resolved modules."""
-        import tempfile
-
-        with tempfile.NamedTemporaryFile(
-            mode="w", suffix=".vera", delete=False
-        ) as f:
-            f.write(source)
-            f.flush()
-            path = f.name
-
-        tree = parse_file(path)
-        ast = transform(tree)
-        return compile(
-            ast, source=source, file=path, resolved_modules=modules,
-        )
-
-    @classmethod
-    def _run_mod(
-        cls, source: str, modules: list,
-        fn: str | None = None, args: list[int] | None = None,
-    ) -> int:
-        """Compile with modules, execute, and return the integer result."""
-        result = cls._compile_mod(source, modules)
-        errors = [d for d in result.diagnostics if d.severity == "error"]
-        assert not errors, f"Unexpected errors: {[e.description for e in errors]}"
-        exec_result = execute(result, fn_name=fn, args=args)
-        assert exec_result.value is not None, "Expected a return value"
-        return exec_result.value
-
-    # -- Basic compilation --------------------------------------------------
-
-    def test_imported_function_compiles(self) -> None:
-        """Imported function produces valid WASM."""
-        mod = self._resolved(("math",), self.MATH_MODULE)
-        result = self._compile_mod("""\
-import math(abs);
-public fn wrap(@Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{ abs(@Int.0) }
-""", [mod])
-        assert result.ok, [d.description for d in result.diagnostics]
-        assert "$abs" in result.wat
-
-    def test_imported_function_executes(self) -> None:
-        """abs(-5) returns 5 via cross-module call."""
-        mod = self._resolved(("math",), self.MATH_MODULE)
-        val = self._run_mod("""\
-import math(abs);
-public fn wrap(@Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{ abs(@Int.0) }
-""", [mod], fn="wrap", args=[-5])
-        assert val == 5
-
-    def test_multiple_imports_execute(self) -> None:
-        """abs(max(x, y)) compiles and runs correctly."""
-        mod = self._resolved(("math",), self.MATH_MODULE)
-        val = self._run_mod("""\
-import math(abs, max);
-public fn abs_max(@Int, @Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{ abs(max(@Int.0, @Int.1)) }
-""", [mod], fn="abs_max", args=[-3, -5])
-        assert val == 3  # abs(max(-3, -5)) = abs(-3) = 3
-
-    # -- Export / visibility -------------------------------------------------
-
-    def test_imported_functions_not_exported(self) -> None:
-        """Imported functions are internal, not WASM exports."""
-        mod = self._resolved(("math",), self.MATH_MODULE)
-        result = self._compile_mod("""\
-import math(abs);
-public fn wrap(@Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{ abs(@Int.0) }
-""", [mod])
-        assert result.ok
-        # Only local public functions are exported
-        assert "wrap" in result.exports
-        assert "abs" not in result.exports
-
-    def test_local_shadows_import(self) -> None:
-        """Local definition of abs shadows the imported one."""
-        mod = self._resolved(("math",), self.MATH_MODULE)
-        val = self._run_mod("""\
-import math(abs);
-public fn abs(@Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{ 999 }
-public fn main(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ abs(42) }
-""", [mod], fn="main")
-        assert val == 999  # local abs, not imported
-
-    # -- Guard rail ----------------------------------------------------------
-
-    def test_guard_rail_still_catches_unknowns(self) -> None:
-        """Unknown function still produces an error even with modules."""
-        mod = self._resolved(("math",), self.MATH_MODULE)
-        result = self._compile_mod("""\
-import math(abs);
-public fn f(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ totally_undefined(1) }
-""", [mod])
-        errors = [d for d in result.diagnostics if d.severity == "error"]
-        assert len(errors) == 1
-        assert "totally_undefined" in errors[0].description
-        assert result.ok is False
-
-    # -- Private helper compilation ------------------------------------------
-
-    def test_private_helper_compiled(self) -> None:
-        """Public fn calling private helper works across modules."""
-        mod = self._resolved(("util",), self.HELPER_MODULE)
-        val = self._run_mod("""\
-import util(double);
-public fn main(@Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{ double(@Int.0) }
-""", [mod], fn="main", args=[7])
-        assert val == 14  # double(7) = internal(7) + internal(7) = 14
-
-    # -- Data imports --------------------------------------------------------
-
-    def test_data_imports_dont_break_codegen(self) -> None:
-        """Importing data types alongside functions compiles fine."""
-        data_mod_source = """\
-public data Color { Red, Green, Blue }
-public fn pick(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ 42 }
-"""
-        mod = self._resolved(("colors",), data_mod_source)
-        val = self._run_mod("""\
-import colors(pick);
-public fn main(-> @Int)
-  requires(true) ensures(true) effects(pure)
-{ pick() }
-""", [mod], fn="main")
-        assert val == 42
-
-
-# =====================================================================
-# Contract failure messages (#112)
-# =====================================================================
-
-class TestContractFailMessages:
-    """Verify that contract violations produce informative error messages."""
-
-    def test_precondition_wat_has_contract_fail(self) -> None:
-        """WAT should call contract_fail before unreachable for requires."""
-        source = """\
-public fn positive(@Int -> @Int)
-  requires(@Int.0 > 0)
-  ensures(true)
-  effects(pure)
-{ @Int.0 }
-"""
-        result = _compile_ok(source)
-        assert "call $vera.contract_fail" in result.wat
-        assert "unreachable" in result.wat
-
-    def test_postcondition_wat_has_contract_fail(self) -> None:
-        """WAT should call contract_fail before unreachable for ensures."""
-        source = """\
-public fn negate(@Int -> @Int)
-  requires(true)
-  ensures(@Int.result > 0)
-  effects(pure)
-{ -@Int.0 }
-"""
-        result = _compile_ok(source)
-        assert "call $vera.contract_fail" in result.wat
-
-    def test_trivial_contract_no_contract_fail(self) -> None:
-        """Trivial contracts should not generate contract_fail calls."""
-        source = """\
-public fn f(@Int -> @Int)
-  requires(true) ensures(true) effects(pure)
-{ @Int.0 }
-"""
-        result = _compile_ok(source)
-        assert "contract_fail" not in result.wat
-
-    def test_precondition_violation_message(self) -> None:
-        """Precondition violation should produce an informative error."""
-        source = """\
-public fn positive(@Int -> @Int)
-  requires(@Int.0 > 0)
-  ensures(true)
-  effects(pure)
-{ @Int.0 }
-"""
-        result = _compile_ok(source)
-        with pytest.raises(RuntimeError, match="Precondition violation"):
-            execute(result, fn_name="positive", args=[0])
-
-    def test_postcondition_violation_message(self) -> None:
-        """Postcondition violation should produce an informative error."""
-        source = """\
-public fn negate(@Int -> @Int)
-  requires(true)
-  ensures(@Int.result > 0)
-  effects(pure)
-{ -@Int.0 }
-"""
-        result = _compile_ok(source)
-        with pytest.raises(RuntimeError, match="Postcondition violation"):
-            execute(result, fn_name="negate", args=[5])
-
-    def test_violation_includes_function_name(self) -> None:
-        """Error message should include the function name."""
-        source = """\
-public fn safe_div(@Int, @Int -> @Int)
-  requires(@Int.0 != 0)
-  ensures(true)
-  effects(pure)
-{ @Int.1 / @Int.0 }
-"""
-        result = _compile_ok(source)
-        with pytest.raises(RuntimeError, match="safe_div"):
-            execute(result, fn_name="safe_div", args=[10, 0])
-
-    def test_violation_includes_contract_text(self) -> None:
-        """Error message should include the contract expression."""
-        source = """\
-public fn bounded(@Int -> @Int)
-  requires(@Int.0 >= 0)
-  requires(@Int.0 <= 100)
-  ensures(true)
-  effects(pure)
-{ @Int.0 }
-"""
-        result = _compile_ok(source)
-        with pytest.raises(RuntimeError, match=r"requires\(@Int.0 >= 0\)"):
-            execute(result, fn_name="bounded", args=[-1])
-
-    def test_postcondition_includes_ensures_text(self) -> None:
-        """Postcondition message should include the ensures expression."""
-        source = """\
-public fn double(@Int -> @Int)
-  requires(true)
-  ensures(@Int.result >= 0)
-  effects(pure)
-{ @Int.0 * 2 }
-"""
-        result = _compile_ok(source)
-        with pytest.raises(RuntimeError, match=r"ensures\(@Int.result >= 0\)"):
-            execute(result, fn_name="double", args=[-1])
-
-    def test_precondition_full_message_format(self) -> None:
-        """Full message should have kind, signature, and clause."""
-        source = """\
-public fn clamp(@Int, @Int, @Int -> @Int)
-  requires(@Int.0 <= @Int.1)
-  ensures(true)
-  effects(pure)
-{ @Int.2 }
-"""
-        result = _compile_ok(source)
-        try:
-            execute(result, fn_name="clamp", args=[5, 3, 4])
-            assert False, "Expected RuntimeError"
-        except RuntimeError as exc:
-            msg = str(exc)
-            assert "Precondition violation" in msg
-            assert "clamp" in msg
-            assert "requires(" in msg
-            assert "@Int.0 <= @Int.1" in msg
-            assert "failed" in msg
-
-    def test_unit_postcondition_violation(self) -> None:
-        """Postcondition on Unit-returning function should report correctly."""
-        source = """\
-public fn bad_put(@Int -> @Unit)
-  requires(true)
-  ensures(new(State<Int>) == old(State<Int>) + 1)
-  effects(<State<Int>>)
-{
-  let @Int = get(());
-  put(@Int.0 + 2);
-  ()
-}
-"""
-        result = _compile_ok(source)
-        with pytest.raises(RuntimeError, match="Postcondition violation"):
-            execute(result, fn_name="bad_put", args=[0],
-                    initial_state={"State_Int": 5})
 
 
 class TestFormatExpr:
