+++ tests/test_codegen_modules.py
@@ -1,0 +1,349 @@
+"""Tests for vera.codegen — Cross-module codegen.
+
+Covers the cross-module guard rail and cross-module function compilation
+via flattening.
+"""
+
+from __future__ import annotations
+
+import pytest
+import wasmtime
+
+from vera.codegen import (
+    CompileResult,
+    ConstructorLayout,
+    ExecuteResult,
+    _align_up,
+    _wasm_type_align,
+    _wasm_type_size,
+    compile,
+    execute,
+)
+from vera.parser import parse_file
+from vera.transform import transform
+
+
+# =====================================================================
+# Helpers
+# =====================================================================
+
+
+def _compile(source: str) -> CompileResult:
+    """Compile a Vera source string to WASM."""
+    # Write to a temp source and parse
+    import tempfile
+    from pathlib import Path
+
+    with tempfile.NamedTemporaryFile(
+        mode="w", suffix=".vera", delete=False
+    ) as f:
+        f.write(source)
+        f.flush()
+        path = f.name
+
+    tree = parse_file(path)
+    ast = transform(tree)
+    return compile(ast, source=source, file=path)
+
+
+def _compile_ok(source: str) -> CompileResult:
+    """Compile and assert no errors."""
+    result = _compile(source)
+    errors = [d for d in result.diagnostics if d.severity == "error"]
+    assert not errors, f"Unexpected errors: {errors}"
+    return result
+
+
+def _run(source: str, fn: str | None = None, args: list[int] | None = None) -> int:
+    """Compile, execute, and return the integer result."""
+    result = _compile_ok(source)
+    exec_result = execute(result, fn_name=fn, args=args)
+    assert exec_result.value is not None, "Expected a return value"
+    return exec_result.value
+
+
+def _run_float(
+    source: str, fn: str | None = None, args: list[int | float] | None = None
+) -> float:
+    """Compile, execute, and return the float result."""
+    result = _compile_ok(source)
+    exec_result = execute(result, fn_name=fn, args=args)
+    assert exec_result.value is not None, "Expected a return value"
+    assert isinstance(exec_result.value, float), (
+        f"Expected float, got {type(exec_result.value).__name__}"
+    )
+    return exec_result.value
+
+
+def _run_io(
+    source: str, fn: str | None = None, args: list[int] | None = None
+) -> str:
+    """Compile, execute, and return captured stdout."""
+    result = _compile_ok(source)
+    exec_result = execute(result, fn_name=fn, args=args)
+    return exec_result.stdout
+
+
+def _run_trap(
+    source: str, fn: str | None = None, args: list[int] | None = None
+) -> None:
+    """Compile, execute, and assert a WASM trap."""
+    result = _compile_ok(source)
+    with pytest.raises((wasmtime.WasmtimeError, wasmtime.Trap, RuntimeError)):
+        execute(result, fn_name=fn, args=args)
+
+
+# =====================================================================
+# Cross-module guard rail
+# =====================================================================
+
+
+class TestCrossModuleGuardRail:
+    """Calls to undefined functions produce a proper diagnostic."""
+
+    def test_undefined_fn_call_diagnostic(self) -> None:
+        """Calling a function not defined in this module emits a diagnostic."""
+        source = """\
+public fn f(-> @Int)
+  requires(true) ensures(true) effects(pure)
+{ unknown_fn(42) }
+"""
+        result = _compile(source)
+        errors = [d for d in result.diagnostics if d.severity == "error"]
+        assert len(errors) == 1
+        assert "unknown_fn" in errors[0].description
+        assert "not defined in this module" in errors[0].description
+        assert "not found in any imported module" in errors[0].description
+        assert result.ok is False
+
+    def test_undefined_fn_no_raw_wasmtime_error(self) -> None:
+        """No raw WAT compilation error -- guard rail catches it first."""
+        source = """\
+public fn f(-> @Int)
+  requires(true) ensures(true) effects(pure)
+{ missing(1) }
+"""
+        result = _compile(source)
+        errors = [d for d in result.diagnostics if d.severity == "error"]
+        assert all("WAT compilation failed" not in e.description for e in errors)
+
+    def test_locally_defined_fn_compiles(self) -> None:
+        """Calls to locally defined functions still work."""
+        source = """\
+public fn helper(-> @Int) requires(true) ensures(true) effects(pure) { 1 }
+public fn f(-> @Int) requires(true) ensures(true) effects(pure) { helper() }
+"""
+        result = _compile_ok(source)
+        assert result.ok is True
+
+
+# =====================================================================
+# Cross-module codegen (C7e)
+# =====================================================================
+
+
+class TestCrossModuleCodegen:
+    """Imported functions are compiled into the WASM module via flattening."""
+
+    # Reusable module sources
+    MATH_MODULE = """\
+public fn abs(@Int -> @Int)
+  requires(true)
+  ensures(@Int.result >= 0)
+  effects(pure)
+{ if @Int.0 < 0 then { 0 - @Int.0 } else { @Int.0 } }
+
+public fn max(@Int, @Int -> @Int)
+  requires(true)
+  ensures(@Int.result >= @Int.0)
+  ensures(@Int.result >= @Int.1)
+  effects(pure)
+{ if @Int.0 >= @Int.1 then { @Int.0 } else { @Int.1 } }
+"""
+
+    HELPER_MODULE = """\
+public fn double(@Int -> @Int)
+  requires(true)
+  ensures(true)
+  effects(pure)
+{ internal(@Int.0) + internal(@Int.0) }
+
+private fn internal(@Int -> @Int)
+  requires(true)
+  ensures(true)
+  effects(pure)
+{ @Int.0 }
+"""
+
+    @staticmethod
+    def _resolved(
+        path: tuple[str, ...], source: str,
+    ) -> "ResolvedModule":
+        """Build a ResolvedModule from source text."""
+        import tempfile
+        from pathlib import Path
+
+        from vera.resolver import ResolvedModule as RM
+
+        with tempfile.NamedTemporaryFile(
+            mode="w", suffix=".vera", delete=False
+        ) as f:
+            f.write(source)
+            f.flush()
+            fpath = f.name
+
+        tree = parse_file(fpath)
+        prog = transform(tree)
+        return RM(
+            path=path,
+            file_path=Path(fpath),
+            program=prog,
+            source=source,
+        )
+
+    @classmethod
+    def _compile_mod(
+        cls, source: str, modules: list,
+    ) -> CompileResult:
+        """Compile with resolved modules."""
+        import tempfile
+
+        with tempfile.NamedTemporaryFile(
+            mode="w", suffix=".vera", delete=False
+        ) as f:
+            f.write(source)
+            f.flush()
+            path = f.name
+
+        tree = parse_file(path)
+        ast = transform(tree)
+        return compile(
+            ast, source=source, file=path, resolved_modules=modules,
+        )
+
+    @classmethod
+    def _run_mod(
+        cls, source: str, modules: list,
+        fn: str | None = None, args: list[int] | None = None,
+    ) -> int:
+        """Compile with modules, execute, and return the integer result."""
+        result = cls._compile_mod(source, modules)
+        errors = [d for d in result.diagnostics if d.severity == "error"]
+        assert not errors, f"Unexpected errors: {[e.description for e in errors]}"
+        exec_result = execute(result, fn_name=fn, args=args)
+        assert exec_result.value is not None, "Expected a return value"
+        return exec_result.value
+
+    # -- Basic compilation --------------------------------------------------
+
+    def test_imported_function_compiles(self) -> None:
+        """Imported function produces valid WASM."""
+        mod = self._resolved(("math",), self.MATH_MODULE)
+        result = self._compile_mod("""\
+import math(abs);
+public fn wrap(@Int -> @Int)
+  requires(true) ensures(true) effects(pure)
+{ abs(@Int.0) }
+""", [mod])
+        assert result.ok, [d.description for d in result.diagnostics]
+        assert "$abs" in result.wat
+
+    def test_imported_function_executes(self) -> None:
+        """abs(-5) returns 5 via cross-module call."""
+        mod = self._resolved(("math",), self.MATH_MODULE)
+        val = self._run_mod("""\
+import math(abs);
+public fn wrap(@Int -> @Int)
+  requires(true) ensures(true) effects(pure)
+{ abs(@Int.0) }
+""", [mod], fn="wrap", args=[-5])
+        assert val == 5
+
+    def test_multiple_imports_execute(self) -> None:
+        """abs(max(x, y)) compiles and runs correctly."""
+        mod = self._resolved(("math",), self.MATH_MODULE)
+        val = self._run_mod("""\
+import math(abs, max);
+public fn abs_max(@Int, @Int -> @Int)
+  requires(true) ensures(true) effects(pure)
+{ abs(max(@Int.0, @Int.1)) }
+""", [mod], fn="abs_max", args=[-3, -5])
+        assert val == 3  # abs(max(-3, -5)) = abs(-3) = 3
+
+    # -- Export / visibility -------------------------------------------------
+
+    def test_imported_functions_not_exported(self) -> None:
+        """Imported functions are internal, not WASM exports."""
+        mod = self._resolved(("math",), self.MATH_MODULE)
+        result = self._compile_mod("""\
+import math(abs);
+public fn wrap(@Int -> @Int)
+  requires(true) ensures(true) effects(pure)
+{ abs(@Int.0) }
+""", [mod])
+        assert result.ok
+        # Only local public functions are exported
+        assert "wrap" in result.exports
+        assert "abs" not in result.exports
+
+    def test_local_shadows_import(self) -> None:
+        """Local definition of abs shadows the imported one."""
+        mod = self._resolved(("math",), self.MATH_MODULE)
+        val = self._run_mod("""\
+import math(abs);
+public fn abs(@Int -> @Int)
+  requires(true) ensures(true) effects(pure)
+{ 999 }
+public fn main(-> @Int)
+  requires(true) ensures(true) effects(pure)
+{ abs(42) }
+""", [mod], fn="main")
+        assert val == 999  # local abs, not imported
+
+    # -- Guard rail ----------------------------------------------------------
+
+    def test_guard_rail_still_catches_unknowns(self) -> None:
+        """Unknown function still produces an error even with modules."""
+        mod = self._resolved(("math",), self.MATH_MODULE)
+        result = self._compile_mod("""\
+import math(abs);
+public fn f(-> @Int)
+  requires(true) ensures(true) effects(pure)
+{ totally_undefined(1) }
+""", [mod])
+        errors = [d for d in result.diagnostics if d.severity == "error"]
+        assert len(errors) == 1
+        assert "totally_undefined" in errors[0].description
+        assert result.ok is False
+
+    # -- Private helper compilation ------------------------------------------
+
+    def test_private_helper_compiled(self) -> None:
+        """Public fn calling private helper works across modules."""
+        mod = self._resolved(("util",), self.HELPER_MODULE)
+        val = self._run_mod("""\
+import util(double);
+public fn main(@Int -> @Int)
+  requires(true) ensures(true) effects(pure)
+{ double(@Int.0) }
+""", [mod], fn="main", args=[7])
+        assert val == 14  # double(7) = internal(7) + internal(7) = 14
+
+    # -- Data imports --------------------------------------------------------
+
+    def test_data_imports_dont_break_codegen(self) -> None:
+        """Importing data types alongside functions compiles fine."""
+        data_mod_source = """\
+public data Color { Red, Green, Blue }
+public fn pick(-> @Int)
+  requires(true) ensures(true) effects(pure)
+{ 42 }
+"""
+        mod = self._resolved(("colors",), data_mod_source)
+        val = self._run_mod("""\
+import colors(pick);
+public fn main(-> @Int)
+  requires(true) ensures(true) effects(pure)
+{ pick() }
+""", [mod], fn="main")
+        assert val == 42
